requirements.txt
# requirements
pandas
openai
streamlit
python-dotenv
PyMuPDF
pytesseract
pillow